code/rag-service/requirements.txt
# RAG Service Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & NLP
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
accelerate==0.25.0
bitsandbytes==0.41.3
einops==0.7.0

# Vector Store
faiss-cpu==1.7.4
# For GPU: faiss-gpu==1.7.4

# Graph Database
neo4j==5.14.0

# Document Store
motor==3.3.2  # Async MongoDB

# Async
aiohttp==3.9.1
aiokafka==0.10.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Utils
numpy==1.26.2
scikit-learn==1.3.2
pyyaml==6.0.1
